Printing
Currently scripts in documents are allowed, so we need to sandbox them
so they can't go and read other documents.

We can't put the document in a sandboxed iframe because that breaks
clean pagination and reflow. Currently, meta sandbox is poorly
supported, as are Service Workers (which would also be a heavyweight
solution).

So we open the window from a sandboxed iframe, because no-allow-same-
origin trickles down to the opened window. We then position that iframe
over the print button, because of popup blockers.