Add style selector
The complicated thing about this feature is that in html, it is usual to
use h1 for the document title. On the other hand, in most word
processors the styles go Title, Subtitle, Header 1, etc. While more user
friendly and closer to what we need for odt documents, implementing
Title and Subtitle would necessitate reimplementing `formatBlock` and
reimplementing creating a normal paragraph on enter rather than another
title, so we don't for now.