Don't blur input elements when clearing them
There are two approaches to doing this: calling focus in a setTimeout
after the click event, and calling preventDefault in the click event.

The first approach leads to the keyboard closing and opening again in
Firefox mobile.

The second approach leads to inconsistent keyboard capitalization on the
first character in Chrome mobile.

This commit does both, calling preventDefault and then calling blur()
and focus() in a setTimeout, fixing both problems.